// schematics/src/pipe/schema.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PwaPipeOptionsSchema {
  path?: string;
  restricted?: boolean;
  tsext?: string;
  flat?: boolean;
  module?: string;
  project?: string;
  /**
   * The name of the pipe.
   */
  name?: string;
  /**
   * When true, does not import this pipe into the owning NgModule.
   */
  skipImport?: boolean;
  /**
   * The extension to generate the pipe in. If empty it is generated in core.
   */
  extension?: string;
}